requirements.txt
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cssbeautifier==1.15.4
Django==5.2.3
django-active-link==0.2.2
django-browser-reload==1.18.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-tailwind==4.0.1
djlint==1.36.4
EditorConfig==0.17.1
Faker==37.4.0
idna==3.10
jedi==0.19.2
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
prompt_toolkit==3.0.51
ptpython==3.0.30
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
